% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{benchmark_signif_bars_gg}
\alias{benchmark_signif_bars_gg}
\title{Significant Proteins Stacked Bars (ggplot2)}
\usage{
benchmark_signif_bars_gg(
  classified_df,
  ev,
  species_colors = NULL,
  title = "Significant Proteins by Species and Direction"
)
}
\arguments{
\item{classified_df}{Classified data frame}

\item{ev}{Expected values data frame}

\item{species_colors}{Named vector of colors per species (optional)}

\item{title}{Plot title}
}
\value{
ggplot2 object
}
\description{
Stacked bar chart of significant proteins by species,
faceted by direction (UP/DOWN).
}
\examples{
data(nadia_dia)
de <- process_proteomics(nadia_dia, verbose = FALSE)$DEPs_results
sp <- utils::read.delim(system.file("extdata", "nadia_dia_spikein.tsv.gz",
                                    package = "NADIA"))
de$Species <- sp$PG.OrganismId[match(de$Protein.IDs, sp$PG.ProteinGroups)]
ev <- data.frame(Comparison = rep(c("B-A", "D-A"), each = 2),
                 Species = c("ECOLI", "YEAST"),
                 expected_logFC = c(1, -0.58, 2, -3.3))
bench <- benchmarking_proteomics(de, ev, verbose = FALSE)
benchmark_signif_bars_gg(bench$classified_df, ev)
}
