% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{bm_plot_extended_ranking_bars}
\alias{bm_plot_extended_ranking_bars}
\title{Horizontal bar chart of extended combined ranking}
\usage{
bm_plot_extended_ranking_bars(extended_ranking, title = NULL)
}
\arguments{
\item{extended_ranking}{List from bm_compute_extended_ranking()}

\item{title}{Optional plot title}
}
\value{
ggplot object
}
\description{
Horizontal bar chart of extended combined ranking
}
