% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{bm_plot_extended_ranking_heatmap}
\alias{bm_plot_extended_ranking_heatmap}
\title{Heatmap of extended ranking (11 metrics)}
\usage{
bm_plot_extended_ranking_heatmap(extended_ranking, title = NULL)
}
\arguments{
\item{extended_ranking}{List from bm_compute_extended_ranking()}

\item{title}{Optional plot title}
}
\value{
ggplot object
}
\description{
Heatmap of extended ranking (11 metrics)
}
