% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{bm_plot_ranking_bars_by_comp}
\alias{bm_plot_ranking_bars_by_comp}
\title{Bar chart of final ranks for a single comparison}
\usage{
bm_plot_ranking_bars_by_comp(ranking_by_comp, comparison, title = NULL)
}
\arguments{
\item{ranking_by_comp}{List returned by \code{bm_compute_ranking_by_comparison()}}

\item{comparison}{Character string: which comparison to plot}

\item{title}{Optional plot title}
}
\value{
ggplot2 object
}
\description{
Bar chart of final ranks for a single comparison
}
