% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{build_clustering_matrix}
\alias{build_clustering_matrix}
\title{Build the clustering matrix aggregated by condition}
\usage{
build_clustering_matrix(
  intensity_matrix,
  sample_metadata,
  selected_features,
  condition_order = NULL,
  aggregate = c("median", "mean")
)
}
\arguments{
\item{intensity_matrix}{Intensity matrix (features x samples)}

\item{sample_metadata}{DataFrame with SampleID, Condition}

\item{selected_features}{Vector of features to include}

\item{condition_order}{Condition order (NULL = alphabetical order)}

\item{aggregate}{Aggregation method: "median" or "mean"}
}
\value{
Matrix of aggregated intensities (features x conditions)
}
\description{
Build the clustering matrix aggregated by condition
}
