% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{cluster_profile_highchart}
\alias{cluster_profile_highchart}
\title{Cluster profile plot with Highcharts}
\usage{
cluster_profile_highchart(
  data,
  cluster,
  conditions = NULL,
  min_membership = NULL,
  show_centroid = TRUE,
  centroid_summary = c("mean", "median"),
  cluster_color = NULL,
  line_width = 1,
  line_opacity = 0.4,
  centroid_width = 3,
  title = NULL,
  height = NULL
)
}
\arguments{
\item{data}{Pattern Profiler DataFrame (LONG format)}

\item{cluster}{Number of the cluster to plot}

\item{conditions}{Vector of condition names (order for the X axis)}

\item{min_membership}{Additional membership filter (NULL = no filter)}

\item{show_centroid}{Show the centroid line (default: TRUE)}

\item{centroid_summary}{Method for the centroid: "mean" or "median"}

\item{cluster_color}{Cluster colour (NULL = automatic)}

\item{line_width}{Width of the profile lines (default: 1)}

\item{line_opacity}{Opacity of the lines (default: 0.4)}

\item{centroid_width}{Width of the centroid line (default: 3)}

\item{title}{Custom title (optional)}

\item{height}{Plot height in pixels}
}
\value{
highchart object
}
\description{
Builds an interactive plot showing the expression profiles of the proteins in
a given cluster.
}
\examples{
if (requireNamespace("Mfuzz", quietly = TRUE) &&
    requireNamespace("e1071", quietly = TRUE)) {
  data(nadia_dia)
  res <- process_proteomics(nadia_dia, verbose = FALSE)
  pp <- pattern_profiler_analysis(res$se_proc, res$DEPs_results,
                                  assay_name = "Impseqrob_min",
                                  auto_select_c = FALSE, c = 3,
                                  verbose = FALSE)

  hc <- cluster_profile_highchart(pp$long_output, cluster = 1,
                                  conditions = c("A", "B", "D"))
  print(class(hc))
}

}
