% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{configure_cluster_palette}
\alias{configure_cluster_palette}
\title{Configurar paleta de colores para clusters}
\usage{
configure_cluster_palette(n_clusters, palette = NULL)
}
\arguments{
\item{n_clusters}{Número de clusters}

\item{palette}{Paleta a usar (NULL, "ggsci::nombre", "brewer:nombre")}
}
\value{
Vector de colores nombrado por cluster
}
\description{
Configurar paleta de colores para clusters
}
