% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{create_expression_set}
\alias{create_expression_set}
\title{Create an ExpressionSet for Mfuzz}
\usage{
create_expression_set(mat, feature_info = NULL)
}
\arguments{
\item{mat}{Expression matrix (features x conditions)}

\item{feature_info}{DataFrame with feature information (optional)}
}
\value{
ExpressionSet object
}
\description{
Create an ExpressionSet for Mfuzz
}
