% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{detect_condition_columns}
\alias{detect_condition_columns}
\title{Detectar columnas de condiciones (z-scores)}
\usage{
detect_condition_columns(data)
}
\arguments{
\item{data}{DataFrame de Pattern Profiler}
}
\value{
Vector de nombres de columnas de condiciones
}
\description{
Detectar columnas de condiciones (z-scores)
}
