% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{.amm_index}
\alias{.amm_index}
\title{Average Maximum Membership (AMM)
Promedio de las memberships máximas. Mayor = mejor.}
\usage{
.amm_index(U)
}
\description{
Average Maximum Membership (AMM)
Promedio de las memberships máximas. Mayor = mejor.
}
