% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Correction.R
\name{.bc_validate_batch}
\alias{.bc_validate_batch}
\title{Validate batch column in colData}
\usage{
.bc_validate_batch(se, batch_column)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{batch_column}{Character. Column name in colData}
}
\description{
Validate batch column in colData
}
\keyword{internal}
