% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.BM_CLASSIFIED_COLS}
\alias{.BM_CLASSIFIED_COLS}
\title{Minimum required columns in the combined classified data.frame}
\usage{
.BM_CLASSIFIED_COLS
}
\description{
Minimum required columns in the combined classified data.frame
}
\keyword{internal}
