% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.BM_CONFUSION_COLS}
\alias{.BM_CONFUSION_COLS}
\title{Required columns in the combined confusion data.frame}
\usage{
.BM_CONFUSION_COLS
}
\description{
Required columns in the combined confusion data.frame
}
\keyword{internal}
