% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_export_data}
\alias{.bm_export_data}
\title{Export data.frame to TSV}
\usage{
.bm_export_data(data, filepath)
}
\arguments{
\item{data}{data.frame}

\item{filepath}{Output path}
}
\description{
Export data.frame to TSV
}
\keyword{internal}
