% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_export_gg_plot}
\alias{.bm_export_gg_plot}
\title{Export ggplot2 to PNG}
\usage{
.bm_export_gg_plot(gg, filepath, width = 12, height = 8, dpi = 150)
}
\arguments{
\item{gg}{ggplot2 object}

\item{filepath}{Output path}

\item{width}{Width in inches}

\item{height}{Height in inches}

\item{dpi}{Resolution}
}
\description{
Export ggplot2 to PNG
}
\keyword{internal}
