% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.BM_EXTENDED_METRICS}
\alias{.BM_EXTENDED_METRICS}
\title{Extended metrics for combined ranking (all higher = better)}
\usage{
.BM_EXTENDED_METRICS
}
\description{
Extended metrics for combined ranking (all higher = better)
}
\keyword{internal}
