% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.BM_METRICS_REQUIRED_COLS}
\alias{.BM_METRICS_REQUIRED_COLS}
\title{Required columns in the combined benchmark metrics data.frame}
\usage{
.BM_METRICS_REQUIRED_COLS
}
\description{
Required columns in the combined benchmark metrics data.frame
}
\keyword{internal}
