% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_performance_to_numeric}
\alias{.bm_performance_to_numeric}
\title{Convert Performance (categorical) to numeric score}
\usage{
.bm_performance_to_numeric(x)
}
\arguments{
\item{x}{Character vector of Performance labels}
}
\value{
Numeric vector
}
\description{
Excellent=5, Very Good=4, Good=3, Acceptable=2, Poor=1, NA=0
}
\keyword{internal}
