% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_read_tsv}
\alias{.bm_read_tsv}
\title{Read a TSV file with readr fallback}
\usage{
.bm_read_tsv(path)
}
\arguments{
\item{path}{File path}
}
\value{
data.frame
}
\description{
Read a TSV file with readr fallback
}
\keyword{internal}
