% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.BM_REQUIRED_COLS}
\alias{.BM_REQUIRED_COLS}
\title{Required columns in the combined opdea data.frame}
\usage{
.BM_REQUIRED_COLS
}
\description{
Required columns in the combined opdea data.frame
}
\keyword{internal}
