% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_validate_bench_metrics}
\alias{.bm_validate_bench_metrics}
\title{Validate combined benchmark metrics data.frame}
\usage{
.bm_validate_bench_metrics(df)
}
\arguments{
\item{df}{data.frame to validate}
}
\description{
Validate combined benchmark metrics data.frame
}
\keyword{internal}
