% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_validate_classified}
\alias{.bm_validate_classified}
\title{Validate combined classified data.frame}
\usage{
.bm_validate_classified(df, p_col = "adj.P.Val")
}
\arguments{
\item{df}{data.frame to validate}

\item{p_col}{Name of the p-value column}
}
\description{
Validate combined classified data.frame
}
\keyword{internal}
