% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_validate_confusion}
\alias{.bm_validate_confusion}
\title{Validate combined confusion data.frame}
\usage{
.bm_validate_confusion(df)
}
\arguments{
\item{df}{data.frame to validate}
}
\description{
Validate combined confusion data.frame
}
\keyword{internal}
