% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{.bm_validate_opdea}
\alias{.bm_validate_opdea}
\title{Validate combined opdea data.frame}
\usage{
.bm_validate_opdea(df)
}
\arguments{
\item{df}{data.frame to validate}
}
\description{
Validate combined opdea data.frame
}
\keyword{internal}
