% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.classify_all_comparisons}
\alias{.classify_all_comparisons}
\title{Classify proteins for all comparisons}
\usage{
.classify_all_comparisons(de_res, ev, alpha, lfc_thr, p_col)
}
\arguments{
\item{de_res}{Data frame with DE results}

\item{ev}{Data frame with expected values}

\item{alpha}{Significance threshold}

\item{lfc_thr}{Log fold-change threshold}

\item{p_col}{P-value column name}
}
\value{
Data frame with classification columns for all comparisons
}
\description{
Classify proteins for all comparisons
}
\keyword{internal}
