% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.compute_auc}
\alias{.compute_auc}
\title{Compute AUC using pROC (optional)}
\usage{
.compute_auc(classified_df, p_col)
}
\arguments{
\item{classified_df}{Classified data frame for one comparison}

\item{p_col}{P-value column name}
}
\value{
AUC value or NA if pROC not available
}
\description{
Compute AUC using pROC (optional)
}
\keyword{internal}
