% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.configure_species_colors}
\alias{.configure_species_colors}
\title{Configure species colors}
\usage{
.configure_species_colors(species, colors = NULL)
}
\arguments{
\item{species}{Character vector of unique species names}

\item{colors}{Named vector of custom colors (optional)}
}
\value{
Named vector of colors per species
}
\description{
Configure species colors
}
\keyword{internal}
