% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.confusion_by_species}
\alias{.confusion_by_species}
\title{Compute confusion matrix by species}
\usage{
.confusion_by_species(classified_df)
}
\arguments{
\item{classified_df}{Classified data frame}
}
\value{
Data frame with TP/FP/TN/FN counts per Comparison x Species
}
\description{
Compute confusion matrix by species
}
\keyword{internal}
