% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.detect_p_col}
\alias{.detect_p_col}
\title{Auto-detect p-value column}
\usage{
.detect_p_col(df, preferred = "adj.P.Val")
}
\arguments{
\item{df}{Data frame}

\item{preferred}{Preferred column name}
}
\value{
Name of the p-value column found
}
\description{
Auto-detect p-value column
}
\keyword{internal}
