% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.export_benchmark_data}
\alias{.export_benchmark_data}
\title{Export benchmark data to TSV or Parquet}
\usage{
.export_benchmark_data(data, filepath, format = "tsv")
}
\arguments{
\item{data}{Data frame to export}

\item{filepath}{Full file path (with extension)}

\item{format}{"tsv" or "parquet"}
}
\description{
Export benchmark data to TSV or Parquet
}
\keyword{internal}
