% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.export_gg_plot}
\alias{.export_gg_plot}
\title{Export ggplot2 chart to PNG}
\usage{
.export_gg_plot(gg, filepath, width = 10, height = 7, dpi = 300)
}
\arguments{
\item{gg}{ggplot2 object}

\item{filepath}{Full file path}

\item{width}{Width in inches (default: 10)}

\item{height}{Height in inches (default: 7)}

\item{dpi}{Resolution (default: 300)}
}
\description{
Export ggplot2 chart to PNG
}
\keyword{internal}
