% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{.extract_limma_results}
\alias{.extract_limma_results}
\title{Extract results from limma fit}
\usage{
.extract_limma_results(
  fit,
  comparisons,
  logFC_up = 1,
  logFC_down = -1,
  alpha = 0.05,
  p_adj = TRUE
)
}
\arguments{
\item{fit}{limma fit object}

\item{comparisons}{Comparison vector}

\item{logFC_up}{Upper logFC threshold for "Up"}

\item{logFC_down}{Lower logFC threshold for "Down"}

\item{alpha}{Significance threshold}

\item{p_adj}{Use adjusted p-value (TRUE) or raw (FALSE)}
}
\value{
Data frame with results
}
\description{
Extract results from limma fit
}
\keyword{internal}
