% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{.fpc_index}
\alias{.fpc_index}
\title{Fuzzy Partition Coefficient (FPC)
Mide claridad del clustering. Mayor = mejor.}
\usage{
.fpc_index(U)
}
\description{
Fuzzy Partition Coefficient (FPC)
Mide claridad del clustering. Mayor = mejor.
}
