% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{.get_NA_overview}
\alias{.get_NA_overview}
\title{NA overview for a SummarizedExperiment}
\usage{
.get_NA_overview(se, assay_name = "log2")
}
\arguments{
\item{se}{SummarizedExperiment}

\item{assay_name}{Assay name to evaluate (default: "log2")}
}
\value{
Data frame with NA statistics per sample
}
\description{
NA overview for a SummarizedExperiment
}
\keyword{internal}
