% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Metrics.R
\name{.im_prone_colors}
\alias{.im_prone_colors}
\title{PRONE-style qualitative color palette}
\usage{
.im_prone_colors(n)
}
\arguments{
\item{n}{Number of colors}
}
\value{
Character vector of hex colors
}
\description{
PRONE-style qualitative color palette
}
\keyword{internal}
