% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_bpca}
\alias{.imp_bpca}
\title{bpca: Bayesian PCA imputation (pcaMethods)}
\usage{
.imp_bpca(x, args = list())
}
\arguments{
\item{args}{list with optional \code{nPcs} (default ncol-1), \code{maxSteps} (default 100)}
}
\description{
bpca: Bayesian PCA imputation (pcaMethods)
}
\keyword{internal}
