% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_Impseqrob}
\alias{.imp_Impseqrob}
\title{Impseqrob: Robust sequential imputation (rrcovNA)}
\usage{
.imp_Impseqrob(x, args = list())
}
\arguments{
\item{args}{list with optional \code{alpha} (default 0.9)}
}
\description{
Impseqrob: Robust sequential imputation (rrcovNA)
}
\keyword{internal}
