% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_knn}
\alias{.imp_knn}
\title{knn: k-nearest neighbors imputation (impute)}
\usage{
.imp_knn(x, args = list())
}
\arguments{
\item{args}{list with optional \code{k} (default 10), \code{rowmax} (default 0.99),
\code{colmax} (default 0.99).}
}
\description{
knn: k-nearest neighbors imputation (impute)
}
\keyword{internal}
