% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_MinDet}
\alias{.imp_MinDet}
\title{MinDet: per-column quantile imputation}
\usage{
.imp_MinDet(x, args = list())
}
\arguments{
\item{args}{list with optional \code{q} (default 0.01)}
}
\description{
MinDet: per-column quantile imputation
}
\keyword{internal}
