% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_MinProb}
\alias{.imp_MinProb}
\title{MinProb: Minimum Probability imputation (imputeLCMD)}
\usage{
.imp_MinProb(x, args = list())
}
\arguments{
\item{args}{list with optional \code{q} (default 0.01), \code{tune.sigma} (default 1)}
}
\description{
MinProb: Minimum Probability imputation (imputeLCMD)
}
\keyword{internal}
