% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_PI}
\alias{.imp_PI}
\title{PI: Perseus-style imputation (down-shifted normal distribution)}
\usage{
.imp_PI(x, args = list())
}
\arguments{
\item{args}{list with optional \code{width} (default 0.3), \code{downshift} (default 1.8)
and \code{seed} (default 1234, para reproducibilidad de rnorm).}
}
\description{
PI: Perseus-style imputation (down-shifted normal distribution)
}
\keyword{internal}
