% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_QRILC}
\alias{.imp_QRILC}
\title{QRILC: Quantile Regression Imputation of Left-Censored data (imputeLCMD)}
\usage{
.imp_QRILC(x, args = list())
}
\arguments{
\item{args}{list with optional \code{tune.sigma} (default 1)}
}
\description{
QRILC: Quantile Regression Imputation of Left-Censored data (imputeLCMD)
}
\keyword{internal}
