% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_with}
\alias{.imp_with}
\title{with: replace NAs with user-specified constant}
\usage{
.imp_with(x, args = list())
}
\arguments{
\item{args}{list with \code{with_value}}
}
\description{
with: replace NAs with user-specified constant
}
\keyword{internal}
