% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{.load_proteomics_data}
\alias{.load_proteomics_data}
\title{Create SummarizedExperiment from proteomics data}
\usage{
.load_proteomics_data(
  data,
  metadata,
  protein_column = "ProteinGroups",
  gene_column = "GeneNames",
  condition_column = "Condition",
  label_column = "Column"
)
}
\arguments{
\item{data}{Data frame with proteins and intensity values}

\item{metadata}{Data frame with sample information}

\item{protein_column}{Protein ID column name}

\item{gene_column}{Gene name column name}

\item{condition_column}{Condition column name in metadata}

\item{label_column}{Sample label column name in metadata}
}
\value{
SummarizedExperiment with assays: raw, log2
}
\description{
Create SummarizedExperiment from proteomics data
}
\keyword{internal}
