% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{.perform_limma}
\alias{.perform_limma}
\title{Run limma analysis}
\usage{
.perform_limma(
  data,
  condition_vector,
  comparisons,
  covariate = NULL,
  block = NULL,
  eBayes_trend = TRUE,
  eBayes_robust = TRUE
)
}
\arguments{
\item{data}{Log2 intensity matrix}

\item{condition_vector}{Condition vector aligned with columns}

\item{comparisons}{Comparison vector}

\item{covariate}{Optional covariate for the model}

\item{eBayes_trend}{Use trend estimation in eBayes (default: TRUE)}

\item{eBayes_robust}{Use robust estimation in eBayes (default: TRUE)}
}
\value{
limma fit object
}
\description{
Run limma analysis
}
\keyword{internal}
