% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{.prepare_pca_input}
\alias{.prepare_pca_input}
\title{Prepare PCA input with differential expression information}
\usage{
.prepare_pca_input(se, DEPs_results, assay_name, alpha = 0.05)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{DEPs_results}{Data frame with DE results}

\item{assay_name}{Assay name to use}

\item{alpha}{Significance threshold for sig_any}
}
\value{
Data frame with intensities and adjP columns per comparison
}
\description{
Prepare PCA input with differential expression information
}
\keyword{internal}
