% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Correction.R
\name{.pvca_categorize}
\alias{.pvca_categorize}
\title{Categorize PVCA labels as technical, biological, interaction, or residual}
\usage{
.pvca_categorize(
  pvca_df,
  technical_factors,
  biological_factors,
  variance_threshold = 0.01
)
}
\description{
Categorize PVCA labels as technical, biological, interaction, or residual
}
\keyword{internal}
