% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Correction.R
\name{.pvca_export_gg_plot}
\alias{.pvca_export_gg_plot}
\title{Export a ggplot2 plot to file}
\usage{
.pvca_export_gg_plot(gg, filepath, width = 10, height = 6, dpi = 150)
}
\description{
Export a ggplot2 plot to file
}
\keyword{internal}
