% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Correction.R
\name{.pvca_prepare_matrix}
\alias{.pvca_prepare_matrix}
\title{Extract and prepare the abundance matrix, handling NAs}
\usage{
.pvca_prepare_matrix(
  se,
  assay_name,
  na_action = "complete",
  fill_value = -1,
  verbose = TRUE
)
}
\description{
Extract and prepare the abundance matrix, handling NAs
}
\keyword{internal}
