% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Correction.R
\name{.pvca_validate_factors}
\alias{.pvca_validate_factors}
\title{Validate that factors exist in colData}
\usage{
.pvca_validate_factors(se, factors)
}
\description{
Validate that factors exist in colData
}
\keyword{internal}
