% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.rename_rownames_from_rd}
\alias{.rename_rownames_from_rd}
\title{Rename rownames using IDs column}
\usage{
.rename_rownames_from_rd(x_df, rd, id_col = "IDs")
}
\arguments{
\item{x_df}{Matrix or data frame with rownames as ProteinGroups}

\item{rd}{rowData data frame with IDs column}

\item{id_col}{Name of IDs column (default: "IDs")}
}
\value{
Data frame with renamed rownames
}
\description{
Rename rownames using IDs column
}
\keyword{internal}
