% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{.se_to_long}
\alias{.se_to_long}
\title{Convert SummarizedExperiment to long format}
\usage{
.se_to_long(se, assay_names = NULL)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{assay_names}{Assay names to include. If NULL, uses all}
}
\value{
Data frame in long format
}
\description{
Convert SummarizedExperiment to long format
}
\keyword{internal}
