% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.sigmoid}
\alias{.sigmoid}
\title{Sigmoid function}
\usage{
.sigmoid(x, k, x0)
}
\arguments{
\item{x}{Numeric vector}

\item{k}{Steepness parameter}

\item{x0}{Midpoint (inflection point)}
}
\value{
Numeric vector in (0, 1)
}
\description{
Sigmoid function
}
\keyword{internal}
