% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEAnalysis.R
\name{.specify_comparisons}
\alias{.specify_comparisons}
\title{Generate comparisons for differential analysis}
\usage{
.specify_comparisons(se, condition_column = NULL, control = NULL)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{condition_column}{Condition column (if NULL, uses SE metadata)}

\item{control}{Control condition. If NULL, generates all pairwise comparisons}
}
\value{
Factor with comparisons in "Treatment-Control" format
}
\description{
Generate comparisons for differential analysis
}
\keyword{internal}
