% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.validate_de_res}
\alias{.validate_de_res}
\title{Validate required columns in DE results}
\usage{
.validate_de_res(de_res)
}
\arguments{
\item{de_res}{Data frame with differential expression results}
}
\value{
Invisible TRUE if valid, stops with error otherwise
}
\description{
Validate required columns in DE results
}
\keyword{internal}
