% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{.validate_expected_values}
\alias{.validate_expected_values}
\title{Validate required columns in expected_values}
\usage{
.validate_expected_values(ev)
}
\arguments{
\item{ev}{Data frame with expected values}
}
\value{
Invisible TRUE if valid, stops with error otherwise
}
\description{
Validate required columns in expected_values
}
\keyword{internal}
