% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{.xie_beni_index}
\alias{.xie_beni_index}
\title{Índice Xie-Beni
Mide compactness vs separación. Menor = mejor.}
\usage{
.xie_beni_index(X, U, centers, m = 2)
}
\description{
Índice Xie-Beni
Mide compactness vs separación. Menor = mejor.
}
