% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{evaluate_cluster_range}
\alias{evaluate_cluster_range}
\title{Evaluate a range of cluster numbers}
\usage{
evaluate_cluster_range(
  eset_std,
  c_range = 2:10,
  m = NULL,
  seeds = c(42, 123, 456),
  verbose = TRUE
)
}
\arguments{
\item{eset_std}{Standardised ExpressionSet}

\item{c_range}{Vector of cluster numbers to evaluate}

\item{m}{Fuzziness parameter}

\item{seeds}{Seeds for reproducibility}

\item{verbose}{Show progress}
}
\value{
DataFrame with the metrics per number of clusters
}
\description{
Evaluate a range of cluster numbers
}
