% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_reuse.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data. Used in mice.reuse.}
\usage{
fetch_data()
}
\value{
data.frame
   the original, non-imputed dataset of the mids object
}
\description{
{ Retrieve the main imputation object when within the
`mice:::sampler` post-imputation calling environment
and return the data object (including missingness)
stored within.
}
}
