% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_tidyHeatmap.R
\name{get_annotation_palette}
\alias{get_annotation_palette}
\title{Build the Colour Palette for Categorical Annotations}
\usage{
get_annotation_palette(levels, palette = NULL)
}
\arguments{
\item{levels}{Vector of levels (categories)}

\item{palette}{Palette specification (same format as get_heatmap_palette)}
}
\value{
Named vector of colours
}
\description{
Build the Colour Palette for Categorical Annotations
}
