% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Metrics.R
\name{im_plot_acc_oi}
\alias{im_plot_acc_oi}
\title{ACC_OI bar plot per imputation method}
\usage{
im_plot_acc_oi(metrics_df, ...)
}
\arguments{
\item{metrics_df}{data.frame from \code{im_compute_metrics()}.}

\item{...}{Additional arguments (unused).}
}
\value{
ggplot object.
}
\description{
Bar chart of Average Correlation Coefficient (ACC_OI), ordered descending
(higher is better).
}
\examples{
data(nadia_dia)
se <- im_prepare_se(nadia_dia, norm_method = "cycloess", verbose = FALSE)
metrics <- im_compute_metrics(se, methods = c("min", "zero", "knn"),
                              verbose = FALSE)
im_plot_acc_oi(metrics)
}
