% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation_Metrics.R
\name{im_plot_pss}
\alias{im_plot_pss}
\title{PSS bar plot per imputation method}
\usage{
im_plot_pss(metrics_df, ...)
}
\arguments{
\item{metrics_df}{data.frame from \code{im_compute_metrics()}.}

\item{...}{Additional arguments (unused).}
}
\value{
ggplot object.
}
\description{
Bar chart of Procrustes Statistical Shape (PSS), ordered ascending
(lower is better). Requires vegan; NA values shown as 0-height bars.
}
\examples{
data(nadia_dia)
se <- im_prepare_se(nadia_dia, norm_method = "cycloess", verbose = FALSE)
metrics <- im_compute_metrics(se, methods = c("min", "zero", "knn"),
                              verbose = FALSE)
im_plot_pss(metrics)
}
