% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_reuse.R
\name{mids.append}
\alias{mids.append}
\title{Joining mice objects. Used in mice.reuse.}
\usage{
mids.append(x, y)
}
\arguments{
\item{x}{mids object provides both data and specification of imputation procedure}

\item{y}{mids object only data information will be retained in the combined object}
}
\value{
mids object
   a new mids object that contains all of `x` and the additional data in `y`
}
\description{
Append one mids object to another. Both objects are expected to have the same variables.
}
\details{
{  Only the data specific aspects are copied (i.e. $data, $imp, $where,
      $nmis), all other information in `y` is discarded. Therefore, only
      the imputation model of `x` is kept and `y` must not contain missing
      data in variables that did not have missing data in `x` (but the
      reverse is allowed).
}
}
