% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missMDA.reuse.R
\name{missMDA.reuse}
\alias{missMDA.reuse}
\title{missMDA.reuse}
\usage{
missMDA.reuse(
  train_data,
  new_data,
  col_type,
  ncp,
  mean_resample = FALSE,
  random.seed = 123,
  maxiter = 998,
  coeff.ridge = 1,
  threshold = 1e-06,
  method = "Regularized"
)
}
\arguments{
\item{train_data}{data.frame used for treining.}

\item{new_data}{data.frame. Df to impute with column names and without target column.}

\item{col_type}{character vector. Vector containing column type names.}

\item{ncp}{return when the training data set was imputed.}

\item{mean_resample}{If TRUE average is used to initiate algorithm if FALSE random resampling is used. More about this can be found in detalist section.}

\item{random.seed}{random seed.}

\item{maxiter}{maximal number of iteration in algortihm.}

\item{coeff.ridge}{Value use in Regularized method.}

\item{threshold}{threshold for convergence.}

\item{method}{method used in imputation algoritm.}
}
\description{
The function allows the user access to missMDA imputation in the A approach.
}
\details{
Function use the same trick as in mice.reuse (new data are changed in NA in imputation stage and added back after it ). Because missMDA use averages to initiate
interactional algorithm. If we decided to use this approach all imputed data in one column will have the same values to avoid this problem. The function allows user to first
randomly resample existing observation in new data from the training set. This is not entirely in the missMDA algorithm but it is still A approach. Becouse missMDA dont save
crated model training has to be performed in this function.
}
