% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Database_Read.R
\name{print.nadia_db}
\alias{print.nadia_db}
\title{Print a summary of a \code{.nadia} file}
\usage{
\method{print}{nadia_db}(x, ...)
}
\arguments{
\item{x}{A \code{nadia_db} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a summary of a \code{.nadia} file
}
