% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{print.proteomics_result}
\alias{print.proteomics_result}
\title{Print method for proteomics_result}
\usage{
\method{print}{proteomics_result}(x, ...)
}
\arguments{
\item{x}{proteomics_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for proteomics_result
}
