% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{run_mfuzz_clustering}
\alias{run_mfuzz_clustering}
\title{Run Mfuzz clustering}
\usage{
run_mfuzz_clustering(eset_std, c, m, seed = 42)
}
\arguments{
\item{eset_std}{Standardised ExpressionSet}

\item{c}{Number of clusters}

\item{m}{Fuzziness parameter}

\item{seed}{Seed for reproducibility}
}
\value{
Mfuzz clustering object
}
\description{
Run Mfuzz clustering
}
