% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{select_optimal_clusters}
\alias{select_optimal_clusters}
\title{Select the optimal number of clusters}
\usage{
select_optimal_clusters(
  eset_std,
  c_range = 2:10,
  m = NULL,
  method = c("xb", "consensus", "elbow"),
  seeds = c(42, 123, 456),
  verbose = TRUE
)
}
\arguments{
\item{eset_std}{Standardised ExpressionSet}

\item{c_range}{Range of cluster numbers to evaluate}

\item{m}{Fuzziness parameter}

\item{method}{Method: "xb", "consensus", "elbow"}

\item{seeds}{Seeds passed on to \code{evaluate_cluster_range()}}

\item{verbose}{Show progress}
}
\value{
List with optimal_c, metrics, m
}
\description{
Select the optimal number of clusters
}
