% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{standardize_eset}
\alias{standardize_eset}
\title{Standardise an ExpressionSet (row-wise z-score)}
\usage{
standardize_eset(eset)
}
\arguments{
\item{eset}{ExpressionSet object}
}
\value{
Standardised ExpressionSet
}
\description{
Standardise an ExpressionSet (row-wise z-score)
}
