% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{summarize_pattern_profiler}
\alias{summarize_pattern_profiler}
\title{Summary of the Pattern Profiler data}
\usage{
summarize_pattern_profiler(data)
}
\arguments{
\item{data}{Pattern Profiler DataFrame}
}
\value{
List with summary statistics
}
\description{
Summary of the Pattern Profiler data
}
\examples{
if (requireNamespace("Mfuzz", quietly = TRUE) &&
    requireNamespace("e1071", quietly = TRUE)) {
  data(nadia_dia)
  res <- process_proteomics(nadia_dia, verbose = FALSE)
  pp <- pattern_profiler_analysis(res$se_proc, res$DEPs_results,
                                  assay_name = "Impseqrob_min",
                                  auto_select_c = FALSE, c = 3,
                                  verbose = FALSE)

  info <- summarize_pattern_profiler(pp$long_output)
  print(info$cluster_summary)
}

}
