% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_tidyHeatmap.R
\name{wrap_heatmap_with_title}
\alias{wrap_heatmap_with_title}
\title{Wrap a Heatmap Together With Its Title}
\usage{
wrap_heatmap_with_title(hm, title, title_size = 14, title_face = "bold")
}
\arguments{
\item{hm}{tidyHeatmap object (InputHeatmap)}

\item{title}{Heatmap title}

\item{title_size}{Font size of the title}

\item{title_face}{Font face of the title}
}
\value{
proteomics_heatmap object (S3 list)
}
\description{
Wrap a Heatmap Together With Its Title
}
