% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Single.R
\name{benchmark_confusion_gg}
\alias{benchmark_confusion_gg}
\title{Confusion Matrix Heatmap (ggplot2)}
\usage{
benchmark_confusion_gg(
  confusion_df,
  title = "Confusion Matrix by Comparison x Species",
  text_size = 3.5,
  axis_text_size = 10
)
}
\arguments{
\item{confusion_df}{Data frame from .confusion_by_species()}

\item{title}{Plot title}

\item{text_size}{Size of cell text labels}

\item{axis_text_size}{Size of axis text}
}
\value{
ggplot2 object
}
\description{
Heatmap of Comparisons x (TP\%, FP\%, FN\%, TN\%) with white-to-blue gradient.
}
