% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking_Multiple.R
\name{bm_plot_extended_ranking_heatmap}
\alias{bm_plot_extended_ranking_heatmap}
\title{Heatmap of extended ranking (11 metrics)}
\usage{
bm_plot_extended_ranking_heatmap(extended_ranking, title = NULL)
}
\arguments{
\item{extended_ranking}{List from bm_compute_extended_ranking()}

\item{title}{Optional plot title}
}
\value{
ggplot object
}
\description{
Heatmap of extended ranking (11 metrics)
}
\examples{
data(nadia_dia)
sp <- utils::read.delim(system.file("extdata", "nadia_dia_spikein.tsv.gz",
                                    package = "NADIA"))
ev <- data.frame(Comparison = rep(c("B-A", "D-A"), each = 2),
                 Species = c("ECOLI", "YEAST"),
                 expected_logFC = c(1, -0.58, 2, -3.3))
opdea_all <- NULL
bench_all <- NULL
for (nm in c("cycloess", "quantile")) {
  de <- process_proteomics(nadia_dia, norm_method = nm,
                           verbose = FALSE)$DEPs_results
  de$Species <- sp$PG.OrganismId[match(de$Protein.IDs, sp$PG.ProteinGroups)]
  opdea_all <- rbind(opdea_all, cbind(compute_opdea_metrics(de, ev), Assay = nm))
  bench_all <- rbind(bench_all, cbind(compute_benchmark_metrics(de, ev), Assay = nm))
}
ext <- bm_compute_extended_ranking(opdea_all, bench_all)
bm_plot_extended_ranking_heatmap(ext)
}
