% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{build_clustering_matrix}
\alias{build_clustering_matrix}
\title{Construir matriz de clustering agregada por condición}
\usage{
build_clustering_matrix(
  intensity_matrix,
  sample_metadata,
  selected_features,
  condition_order = NULL,
  aggregate = c("median", "mean")
)
}
\arguments{
\item{intensity_matrix}{Matriz de intensidades (features x samples)}

\item{sample_metadata}{DataFrame con SampleID, Condition}

\item{selected_features}{Vector de features a incluir}

\item{condition_order}{Orden de condiciones (NULL = orden alfabético)}

\item{aggregate}{Método de agregación: "median" o "mean"}
}
\value{
Matriz de intensidades agregadas (features x conditions)
}
\description{
Construir matriz de clustering agregada por condición
}
