% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{configure_cluster_palette}
\alias{configure_cluster_palette}
\title{Configure the colour palette for the clusters}
\usage{
configure_cluster_palette(n_clusters, palette = NULL)
}
\arguments{
\item{n_clusters}{Number of clusters}

\item{palette}{Palette to use (NULL, "ggsci::name", "brewer:name")}
}
\value{
Vector of colours named by cluster
}
\description{
Configure the colour palette for the clusters
}
