% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{create_expression_set}
\alias{create_expression_set}
\title{Crear ExpressionSet para Mfuzz}
\usage{
create_expression_set(mat, feature_info = NULL)
}
\arguments{
\item{mat}{Matriz de expresión (features x conditions)}

\item{feature_info}{DataFrame con info de features (opcional)}
}
\value{
Objeto ExpressionSet
}
\description{
Crear ExpressionSet para Mfuzz
}
