% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{detect_condition_columns}
\alias{detect_condition_columns}
\title{Detect the condition columns (z-scores)}
\usage{
detect_condition_columns(data)
}
\arguments{
\item{data}{Pattern Profiler DataFrame}
}
\value{
Vector with the names of the condition columns
}
\description{
Detect the condition columns (z-scores)
}
