% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_missForest}
\alias{.imp_missForest}
\title{missForest: Random Forest imputation}
\usage{
.imp_missForest(x, args = list())
}
\arguments{
\item{args}{list with optional \code{maxiter} (default 10), \code{ntree} (default 100),
\code{mtry} (default floor(nrow(x)^(1/3)), NAguideR cube root strategy)}
}
\description{
missForest: Random Forest imputation
}
\keyword{internal}
