% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{.imp_MLE}
\alias{.imp_MLE}
\title{MLE: Maximum Likelihood Estimation (norm)}
\usage{
.imp_MLE(x, args = list())
}
\arguments{
\item{args}{list with optional \code{seed} (default 123).
NAguideR strategy: no transpose (features x samples).}
}
\description{
MLE: Maximum Likelihood Estimation (norm)
}
\keyword{internal}
