% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{.prepare_metadata}
\alias{.prepare_metadata}
\title{Prepare metadata from proteomics_data object}
\usage{
.prepare_metadata(preprocessing, covariate_df = NULL)
}
\arguments{
\item{preprocessing}{proteomics_data list (output of preprocess_spectronaut or preprocess_tmt)}
}
\value{
Data frame with columns: Column, Condition, Replicate
}
\description{
Prepare metadata from proteomics_data object
}
\keyword{internal}
