% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{.prepare_protein_data}
\alias{.prepare_protein_data}
\title{Prepare protein data from proteomics_data object}
\usage{
.prepare_protein_data(preprocessing)
}
\arguments{
\item{preprocessing}{A \code{proteomics_data} object, from any of the preprocess_*() functions}
}
\value{
Data frame with columns: ProteinGroups, GeneNames, UniqPepts, and intensity columns
}
\description{
Prepare protein data from proteomics_data object
}
\keyword{internal}
