% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{evaluate_cluster_range}
\alias{evaluate_cluster_range}
\title{Evaluar rango de números de clusters}
\usage{
evaluate_cluster_range(
  eset_std,
  c_range = 2:10,
  m = NULL,
  seeds = c(42, 123, 456),
  verbose = TRUE
)
}
\arguments{
\item{eset_std}{ExpressionSet estandarizado}

\item{c_range}{Vector de números de clusters a evaluar}

\item{m}{Parámetro de fuzziness}

\item{seeds}{Semillas para reproducibilidad}

\item{verbose}{Mostrar progreso}
}
\value{
DataFrame con métricas por número de clusters
}
\description{
Evaluar rango de números de clusters
}
