% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{extract_se_data}
\alias{extract_se_data}
\title{Extract data from a SummarizedExperiment}
\usage{
extract_se_data(se_proc, assay_name = NULL)
}
\arguments{
\item{se_proc}{SummarizedExperiment object}

\item{assay_name}{Name of the assay to use (NULL = first available)}
}
\value{
List with:
\itemize{
\item intensity_matrix: features x samples matrix
\item feature_ids: vector of protein IDs
\item sample_metadata: DataFrame with SampleID, Condition, etc.
}
}
\description{
Retrieves the intensity matrix, the feature IDs and the sample metadata from a
SummarizedExperiment object.
}
