% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_tidyHeatmap.R
\name{get_annotation_palette}
\alias{get_annotation_palette}
\title{Obtener paleta de colores para anotaciones categóricas}
\usage{
get_annotation_palette(levels, palette = NULL)
}
\arguments{
\item{levels}{Vector de niveles (categorías)}

\item{palette}{Especificación de paleta (mismo formato que get_heatmap_palette)}
}
\value{
Vector nombrado de colores
}
\description{
Obtener paleta de colores para anotaciones categóricas
}
