% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization_Metrics.R
\name{nm_plot_boxplot}
\alias{nm_plot_boxplot}
\title{Intensity boxplot per method}
\usage{
nm_plot_boxplot(se, assay_names = NULL, condition_col = "Condition", ...)
}
\arguments{
\item{se}{SummarizedExperiment from \code{import_norm_matrices()}.}

\item{assay_names}{Character vector of assay names to include. NULL = all.}

\item{condition_col}{Column in colData with condition labels.}

\item{...}{Additional arguments (unused).}
}
\value{
ggplot object.
}
\description{
Displays distribution of protein intensities for each sample,
faceted by normalization method.
}
\examples{
data(nadia_dia)
se <- nm_prepare_se(nadia_dia, verbose = FALSE)
se <- nm_run_normalizations(se, methods = c("cycloess", "quantile", "MAD"),
                            verbose = FALSE)

nm_plot_boxplot(se)

}
