% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{print.proteomics_result}
\alias{print.proteomics_result}
\title{Print method for proteomics_result}
\usage{
\method{print}{proteomics_result}(x, ...)
}
\arguments{
\item{x}{proteomics_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{x}, invisibly. Called for the summary it prints: the assays of the
SummarizedExperiment, the comparisons performed, the number of differential
proteins and the parameters used.
}
\description{
Print method for proteomics_result
}
\examples{
data(nadia_dia)
res <- process_proteomics(nadia_dia, verbose = FALSE)
print(res)

}
