% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{run_mfuzz_clustering}
\alias{run_mfuzz_clustering}
\title{Ejecutar clustering Mfuzz}
\usage{
run_mfuzz_clustering(eset_std, c, m, seed = 42)
}
\arguments{
\item{eset_std}{ExpressionSet estandarizado}

\item{c}{Número de clusters}

\item{m}{Parámetro de fuzziness}

\item{seed}{Semilla para reproducibilidad}
}
\value{
Objeto de clustering Mfuzz
}
\description{
Ejecutar clustering Mfuzz
}
