% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{select_optimal_clusters}
\alias{select_optimal_clusters}
\title{Seleccionar número óptimo de clusters}
\usage{
select_optimal_clusters(
  eset_std,
  c_range = 2:10,
  m = NULL,
  method = c("xb", "consensus", "elbow"),
  verbose = TRUE
)
}
\arguments{
\item{eset_std}{ExpressionSet estandarizado}

\item{c_range}{Rango de clusters a evaluar}

\item{m}{Parámetro de fuzziness}

\item{method}{Método: "xb", "consensus", "elbow"}

\item{verbose}{Mostrar progreso}
}
\value{
Lista con optimal_c, metrics, m
}
\description{
Seleccionar número óptimo de clusters
}
