% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Analysis.R
\name{standardize_eset}
\alias{standardize_eset}
\title{Estandarizar ExpressionSet (z-score por fila)}
\usage{
standardize_eset(eset)
}
\arguments{
\item{eset}{Objeto ExpressionSet}
}
\value{
ExpressionSet estandarizado
}
\description{
Estandarizar ExpressionSet (z-score por fila)
}
