% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{summarize_pattern_profiler}
\alias{summarize_pattern_profiler}
\title{Resumen de datos de Pattern Profiler}
\usage{
summarize_pattern_profiler(data)
}
\arguments{
\item{data}{DataFrame de Pattern Profiler}
}
\value{
Lista con estadísticas resumidas
}
\description{
Resumen de datos de Pattern Profiler
}
