% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Profiler_Highcharts.R
\name{summarize_pattern_profiler}
\alias{summarize_pattern_profiler}
\title{Summary of the Pattern Profiler data}
\usage{
summarize_pattern_profiler(data)
}
\arguments{
\item{data}{Pattern Profiler DataFrame}
}
\value{
List with summary statistics
}
\description{
Summary of the Pattern Profiler data
}
